backend/requirements.txt
pymongo
flask-jwt-extended
bcrypt
PyPDF2
pymongo

flask
flask-cors
langchain
openai
werkzeug
python-dotenv
langchain-google-genai
docx2txt
PyPDF2
